requirements.txt
Package            Version
------------------ ---------
alembic            0.7.0
amqp               2.6.1
attrs              20.3.0
certifi            2020.12.5
chardet            4.0.0
dnspython          1.16.0
eventlet           0.30.2
greenlet           1.0.0
idna               2.10
importlib-metadata 3.7.0
iniconfig          1.1.1
Jinja2             2.11.3
kombu              4.6.11
Mako               1.1.4
MarkupSafe         1.1.1
mock               4.0.3
nameko             2.13.0
packaging          20.9
path               15.1.2
path.py            12.5.0
pip                21.0.1
pluggy             0.13.1
psycopg2           2.7.7
py                 1.10.0
pyparsing          2.4.7
pytest             6.2.2
PyYAML             5.4.1
redis              3.5.3
requests           2.25.1
setuptools         46.1.3
six                1.15.0
SQLAlchemy         1.4.3
toml               0.10.2
typing-extensions  3.7.4.3
urllib3            1.26.3
uuid               1.30
vine               1.3.0
Werkzeug           1.0.1
wheel              0.34.2
wrapt              1.12.1
zipp               3.4.1
uuid==1.30
